+++ tests/test_model/test_qwen3_omni_mmsu_talker_ci.py
@@ -36,16 +36,13 @@ from tests.test_model.omni_router_utils import (
     ManagedRouterHandle,
     router_worker_traffic_guard,
 )
-from tests.test_model.omni_whisper_wer_utils import (
-    omni_whisper_wer_router,  # noqa: F401 - pytest fixture imported by name.
-    wait_for_gpu_memory_release,
-)
 from tests.utils import (
     MetricCheckCollector,
     apply_slack,
     apply_wer_slack,
     assert_speed_thresholds,
     assert_wer_partitioned,
+    wait_for_gpu_memory_release,
 )
 
 MAX_SAMPLES = 40
