+++ tests/test_model/test_qwen3_omni_tts_ci.py
@@ -33,10 +33,6 @@ from tests.test_model.omni_router_utils import (
     print_worker_snapshot,
     router_get_json,
 )
-from tests.test_model.omni_whisper_wer_utils import (
-    omni_whisper_wer_router,  # noqa: F401 - pytest fixture imported by name.
-    wait_for_gpu_memory_release,
-)
 from tests.utils import (
     MetricCheckCollector,
     apply_slack,
@@ -46,6 +42,7 @@ from tests.utils import (
     assert_summary_metrics,
     assert_wer_partitioned,
     no_proxy_env,
+    wait_for_gpu_memory_release,
 )
 
 PROJECT_ROOT = Path(__file__).resolve().parents[2]
